Split Up Cluster Provisioners
Make applications more abstract by giving them an application name and
scope, so it's handled in one place.  Split the (cluster) provisoners
into their own packages.  Rework the Argo stuff so we can derive and
lookup cluster names, rather than rely on server addresses from configs
that may or may not exist.
