+++ pkg/provisioners/application/provisioner.go
@@ -20,6 +20,7 @@ import (
 	"context"
 	"errors"
 
+	"github.com/eschercloudai/unikorn/pkg/constants"
 	"github.com/eschercloudai/unikorn/pkg/provisioners"
 	"github.com/eschercloudai/unikorn/pkg/readiness"
 
@@ -40,15 +41,23 @@ type Provisioner struct {
 	// client provides access to Kubernetes.
 	client client.Client
 
+	// name is the application name.
+	name string
+
 	// object is the object's required state.
 	object client.Object
+
+	// labels defines a unique application label selector.
+	labels labels.Set
 }
 
 // New returns a new initialized provisioner object.
-func New(client client.Client, object client.Object) *Provisioner {
+func New(client client.Client, name string, scope labels.Set, object client.Object) *Provisioner {
 	return &Provisioner{
 		client: client,
+		name:   name,
 		object: object,
+		labels: labels.Merge(scope, labels.Set{constants.ApplicationLabel: name}),
 	}
 }
 
@@ -75,7 +84,7 @@ func (p *Provisioner) toUnstructured() (*unstructured.Unstructured, error) {
 // generated names here as it's a multi-tenant platform, argo enforces the use of a single
 // namespace, and we want users to be able to define their own names irrespective
 // of other users.
-func (p *Provisioner) findApplication(ctx context.Context, u *unstructured.Unstructured) (*unstructured.Unstructured, error) {
+func (p *Provisioner) findApplication(ctx context.Context) (*unstructured.Unstructured, error) {
 	resources := &unstructured.UnstructuredList{
 		Object: map[string]interface{}{
 			"apiVersion": "argoproj.io/v1alpha1",
@@ -83,7 +92,7 @@ func (p *Provisioner) findApplication(ctx context.Context, u *unstructured.Unstr
 		},
 	}
 
-	selector := labels.SelectorFromSet(u.GetLabels())
+	selector := labels.SelectorFromSet(p.labels)
 
 	if err := p.client.List(ctx, resources, &client.ListOptions{Namespace: "argocd", LabelSelector: selector}); err != nil {
 		return nil, err
@@ -106,7 +115,7 @@ func (p *Provisioner) findApplication(ctx context.Context, u *unstructured.Unstr
 func (p *Provisioner) Provision(ctx context.Context) error {
 	log := log.FromContext(ctx)
 
-	log.Info("provisioning application")
+	log.Info("provisioning application", "application", p.name)
 
 	// Convert the generic object type into unstructured for the next bit...
 	required, err := p.toUnstructured()
@@ -114,16 +123,19 @@ func (p *Provisioner) Provision(ctx context.Context) error {
 		return err
 	}
 
+	required.SetGenerateName(p.name + "-")
+	required.SetLabels(p.labels)
+
 	// Resource, after provisioning, should be set to either the existing resource
 	// or the newly created one.  The point here is the API will have filled in
 	// the name so we can perform readiness checks.
-	resource, err := p.findApplication(ctx, required)
+	resource, err := p.findApplication(ctx)
 	if err != nil {
 		return err
 	}
 
 	if resource == nil {
-		log.Info("creating new application")
+		log.Info("creating new application", "application", p.name)
 
 		if err := p.client.Create(ctx, required); err != nil {
 			return err
@@ -131,7 +143,7 @@ func (p *Provisioner) Provision(ctx context.Context) error {
 
 		resource = required
 	} else {
-		log.Info("updating existing application")
+		log.Info("updating existing application", "application", p.name)
 
 		// Replace the specification with what we expect.
 		temp := resource.DeepCopy()
@@ -142,7 +154,7 @@ func (p *Provisioner) Provision(ctx context.Context) error {
 		}
 	}
 
-	log.Info("waiting for application to become healthy")
+	log.Info("waiting for application to become healthy", "application", p.name)
 
 	applicationHealthy := readiness.NewApplicationHealthy(p.client, resource)
 
@@ -150,7 +162,7 @@ func (p *Provisioner) Provision(ctx context.Context) error {
 		return err
 	}
 
-	log.Info("application provisioned")
+	log.Info("application provisioned", "application", p.name)
 
 	return nil
 }
@@ -159,26 +171,20 @@ func (p *Provisioner) Provision(ctx context.Context) error {
 func (p *Provisioner) Deprovision(ctx context.Context) error {
 	log := log.FromContext(ctx)
 
-	log.Info("deprovisioning application")
+	log.Info("deprovisioning application", "application", p.name)
 
-	// Convert the generic object type into unstructured for the next bit...
-	required, err := p.toUnstructured()
-	if err != nil {
-		return err
-	}
-
-	resource, err := p.findApplication(ctx, required)
+	resource, err := p.findApplication(ctx)
 	if err != nil {
 		return err
 	}
 
 	if resource == nil {
-		log.Info("application does not exist")
+		log.Info("application does not exist", "application", p.name)
 
 		return nil
 	}
 
-	log.Info("adding application finalizer")
+	log.Info("adding application finalizer", "application", p.name)
 
 	// Apply a finalizer to ensure synchronous deletion. See
 	// https://argo-cd.readthedocs.io/en/stable/user-guide/app_deletion/
@@ -189,19 +195,19 @@ func (p *Provisioner) Deprovision(ctx context.Context) error {
 		return err
 	}
 
-	log.Info("deleting application")
+	log.Info("deleting application", "application", p.name)
 
 	if err := p.client.Delete(ctx, resource); err != nil {
 		return err
 	}
 
-	log.Info("waiting for application deletion")
+	log.Info("waiting for application deletion", "application", p.name)
 
 	if err := readiness.NewRetry(readiness.NewResourceNotExists(p.client, resource)).Check(ctx); err != nil {
 		return err
 	}
 
-	log.Info("application deleted")
+	log.Info("application deleted", "application", p.name)
 
 	return nil
 }
